Updated Base.h and Static Linking
Updated Base.h file so it can handle both dynamic and static linking on windows platform, and moved Motus to static linking by default

+++ Motus/src/Platform/Windows/WindowsWindow.cpp
@@ -109,7 +109,7 @@ namespace Motus {
 		glfwSetCursorPosCallback(m_Window, [](GLFWwindow* window, double xpos, double ypos) {
 			WindowData& data = *(WindowData*)glfwGetWindowUserPointer(window);
 
-			MouseMovedEvent event(xpos, ypos);
+			MouseMovedEvent event((uint16_t)xpos, (uint16_t)ypos);
 			data.EventCallback(event);
 			});				// Mouse Moved
 		glfwSetMouseButtonCallback(m_Window, [](GLFWwindow* window, int button, int action, int mods) {
